+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Added
+
+- `responseErrorDetails` when receiving an error response for host asset
+  listing.
+- Logging of host IDs data type.
+
 ## 4.2.2 2020-10-12
 
 ### Changed